user-management/pyproject.toml
[tool.poetry]
name = "user-management"
version = "0.1.0"
description = ""
authors = ["hubaib <hubaibm9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlmodel = "^0.0.19"
psycopg = "^3.1.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
